chore: 划分 interface
