+++ packages/iceworks-server/src/interface/index.ts
@@ -1,0 +1,14 @@
+export * from './project';
+export * from './material';
+
+export interface IProjectsResult {
+  projects: any[];
+}
+export interface IProjectResult {
+  project: any;
+}
+
+export interface IProjectService {
+  getProjects(): Promise<IProjectsResult>;
+  getCurrent(): Promise<IProjectResult>;
+}
