+++ packages/iceworks-server/src/interface/material.ts
@@ -1,0 +1,19 @@
+/**
+ * TODO 物料的模板信息
+ */
+export interface IMaterialScaffold {
+
+}
+
+/**
+ * TODO 物料的区块信息
+ */
+export interface IMaterialBlock {
+
+}
+/**
+ * TODO 物料的组件信息
+ */
+export interface IMaterialComponent {
+
+}
